backend/requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
celery==5.5.3
redis==7.0.1
pymongo==4.15.3
python-dotenv==1.2.1
python-multipart==0.0.20
pillow==12.0.0
google-generativeai==0.8.5
pydantic==2.12.4
broadcaster[redis]